fix styling and redirects
